+++ src/format_wav.h
@@ -11,48 +11,60 @@
 
 namespace spectral {
 namespace format {
+namespace chunk {
 
 
-class wav : public riff {
+class wav_format : public riff::chunk_t {
 public:
-  class subchunk_audio_format : public riff::subchunk {
+  static const dword Tag = ' ' << 24 | 't' << 16 | 'm' << 8 | 'f';  // "fmt "
+  word Format, NumOfChannels, BlockAlign;
+  dword SamplingRate, AverageBPS;
+  std::unique_ptr<byte[]> SpecificFields;
+
+  class factory : public chunk_t::factory {
   public:
-    subchunk_audio_format( int32_t ID, int32_t Size,
-                           int16_t Tag, int16_t NumOfChannels, int16_t BlockAlign,
-                           int32_t SamplingRate, int32_t AverageBPS );
-    subchunk_audio_format( int32_t ID,
-                           int16_t Tag, int16_t NumOfChannels, int16_t BlockAlign,
-                           int32_t SamplingRate, int32_t AverageBPS );
-    virtual ~subchunk_audio_format() = default;
-    virtual void Write( FILE *File ) const override;
-
-  protected:
-    static const int32_t DataSize = sizeof(int16_t) * 3 + sizeof(int32_t) * 2;
-    int16_t Tag, NumOfChannels, BlockAlign;
-    int32_t SamplingRate, AverageBPS;
+    factory() = default;
+    virtual ~factory() = default;
+    virtual std::shared_ptr<chunk_t> Create( dword ID, dword Size ) const override;
   };
+  friend factory;
 
+  virtual ~wav_format() = default;
+  virtual void Clear() override;
+  virtual void Write( FILE *File ) const override;
+  std::ostream & Print( std::ostream &Stream ) const override;
+  virtual void Read( FILE *File ) override;
 
-  class subchunk_audio_data : public riff::subchunk {
-  public:
-    subchunk_audio_data( int32_t NumOfSamples, int32_t SizeOfSample,
-                         int32_t NumOfChannels, void *Data );
-    virtual ~subchunk_audio_data() = default;
-    virtual void Write( FILE *File ) const override;
-
-  protected:
-    static const int32_t Tag = 'a' << 24 | 't' << 16 | 'a' << 8 | 'd';  // "data"
-    int32_t NumOfSamples, SizeOfSample, NumOfChannels;
-    void *Data;
-  };
-
-
-  wav();
-  virtual ~wav() = default;
-  virtual void Read( const std::string &FileName ) override;
-  virtual void Write( const std::string &FileName ) const override;
+protected:
+  static const int CommonFieldSize = sizeof(word) * 3 + sizeof(dword) * 2;
+  wav_format( dword ID, dword Size );
 };
 
 
+class wav_data : public riff::chunk_t {
+public:
+  static const int32_t Tag = 'a' << 24 | 't' << 16 | 'a' << 8 | 'd';  // "data"
+  std::unique_ptr<byte[]> Data;
+
+  class factory : public chunk_t::factory {
+  public:
+    factory() = default;
+    virtual ~factory() = default;
+    virtual std::shared_ptr<chunk_t> Create( dword ID, dword Size ) const override;
+  };
+  friend factory;
+
+  virtual ~wav_data() = default;
+  virtual void Clear() override;
+  virtual void Write( FILE *File ) const override;
+  std::ostream & Print( std::ostream &Stream ) const override;
+  virtual void Read( FILE *File ) override;
+
+protected:
+  wav_data( dword ID, dword Size );
+};
+
+
+} /* End of 'chunk' namespace */
 } /* End of 'format' namespace */
 } /* End of 'spectral' namespace */