+++ src/pch.h
@@ -6,6 +6,7 @@
 /*****
  * STL
  *****/
+#include <iostream>
 #include <memory>
 #include <string>
 #include <vector>