+++ CHANGELOG.md
@@ -2,8 +2,15 @@
 All notable changes to this project will be documented in this file.
 
 
-## [0.1.4] - 2020-12-23
+## [0.3.0] - 2020-12-25
+### Added:
+- WAVE chunk handlers (fmt/data)
+- RIFF data printout
+
+
+## [0.1.4] - 2020-12-25
 ### Added:
 - 'format_t' class
 - 'format::riff' class
-- Default RIFF chunk reader (skip)
+- Default RIFF chunk handler (copy)
+